Move the old "explain" command to "xai"
